Option to represent all C++ operators.
In its default behavior, bindgen does not emit information for any
C++ operatorXYZ function (unless operatorXYZ is a valid identifier,
which it isn't for C++ overloaded operators).

This PR introduces a new command line option to represent all operators.
This is not useful for those who directly consume the output of bindgen,
but is useful for post-processors. Specifically, consumers will need to
implement the callback to rename these items to something more useful.

Part of https://github.com/google/autocxx/issues/124