To Be Squashed
Use the different radix literals for integers that were used in the original C
and C++ header files, now that it is supported. The new values were confirmed
by hand to be the same as the original ones - they are just represented using
the Rust equivalent of the original form.