Don't require a `Box` wrapper on `Bindings::write()`
The function in question doesn't store the `dyn Write` trait object
anywhere, allowing us to pass a plain `&mut dyn Write` (which is also
implemented for `&mut Box<dyn Write>` allowing the existing CLI parser
to operate normally).

This in turn no longer requires users of `bindgen` to wrap their
`&mut File` or `&mut Vec<u8>` in a moved `Box` just to be able to give
temporary mutable access to the `write() function`.